+++ VulkanEngine/talos/gameobjects/Light.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "config.h"
+#include "../config.h"
 
 struct Light {
 	glm::vec3 position;
