+++ VulkanEngine/talos/image/Texture.cpp
@@ -1,7 +1,7 @@
 #include "Texture.h"
-#include "stb_image.h"
-#include "Memory.h"
-#include "Descriptors.h"
+#include "../../stb_image.h"
+#include "../utilities/Memory.h"
+#include "../init/Descriptors.h"
 
 namespace vkImage {
