+++ VulkanEngine/talos/mesh/ObjMesh.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "config.h"
+#include "../config.h"
 
 namespace vkMesh {
 	struct Material {
