+++ VulkanEngine/talos/utilities/ModelLoader.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "config.h"
+#include "../config.h"
 
 namespace talos {
 	namespace util {
